requirements.txt
datasets>=2.0.0
rouge-score>=0.1.2
sacrebleu>=2.0.0
bert-score>=0.3.12
numpy>=1.20.0
scikit-learn>=1.0.0
tqdm>=4.60.0

python-dotenv
openai>=1.0.0
